♻️ spread out and organize frontend
